Explicitly flag failed tests in JSON output.
Updates the GradescopeJsonFormatter to explicitly flag failed tests as
failed. This mainly matters when tests are worth 0 points - if `status`
is not specified in the JSON, then Gradescope assumes that a test that
earned max points has passed. So, when a test is worth 0 points, it
shows as passed even if it fails or aborts. This change explicitly
includes the `status` field in the JSON if a test has been marked as
failing.

See the Test Case Status section in the Gradescope specifications:
https://gradescope-autograders.readthedocs.io/en/latest/specs/
for more detail.